Consolidate Live Activity payload builder into ios.js
Moves createLiveActivityPayload() from the standalone live-activity.js
into ios.js as a second export, keeping all iOS-specific logic in one
place consistent with the existing android.js/ios.js pattern.
